requirements.txt
firebase_admin==6.5.0
pyserial==3.5
rich==13.9.4
